backend/requirements.txt
# Django and REST framework
Django>=5.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-filter>=23.5

# Authentication
django-allauth>=0.60.1
django-rest-auth>=0.9.5
dj-rest-auth>=5.0.2
djangorestframework-simplejwt>=5.3.1
django-otp>=1.2.2
qrcode>=7.4.2

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Word Cloud Generation
wordcloud>=1.9.2
matplotlib>=3.8.2
Pillow>=10.2.0

# OpenAI API
openai>=1.12.0

# Azure Blob Storage
azure-storage-blob>=12.19.0

# Environment variables
python-dotenv>=1.0.1

# API Documentation
drf-yasg>=1.21.7

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
